refactor(YALB-639): namespaced controls
