+++ components/00-tokens/colors/color-pairings.twig
@@ -5,11 +5,11 @@
     <h2>Component Theme: {{ theme }}</h2>
     <p class="text-field">Here's a paragraph with some text. And <a href="#">links can go here</a> too. Isn't that cool?</p>
     <div class="cta-group">
-      {% include "@atoms/controls/cta/cta.twig" with {
+      {% include "@atoms/controls/cta/yds-cta.twig" with {
         cta__content: 'Cta Link',
         cta__href: '#',
       } %}
-      {% include "@atoms/controls/cta/cta.twig" with {
+      {% include "@atoms/controls/cta/yds-cta.twig" with {
         cta__content: 'Outline',
         cta__style: 'outline',
         cta__href: '#',
