+++ components/01-atoms/atoms.scss
@@ -1,6 +1,6 @@
 @forward './controls/button/button' as button-*;
-@forward './controls/cta/cta';
-@forward './controls/text-link/text-link';
+@forward './controls/cta/yds-cta';
+@forward './controls/text-link/yds-text-link';
 @forward './date-time/date-time';
 @forward './divider/divider';
 @forward './images/icons/icons';
