+++ components/01-atoms/controls/cta/yds-cta.twig
@@ -23,7 +23,7 @@
   'data-cta-outline-weight': cta__outline_weight|default('2'),
 }) %}
 
-{% include "@atoms/controls/base/control.twig" with {
+{% include "@atoms/controls/base/yds-control.twig" with {
   control__base_class: cta__base_class,
   control__modifiers: cta__modifiers,
   control__blockname: cta__blockname,
