+++ components/01-atoms/typography/headings/_headings.scss
@@ -1,5 +1,5 @@
 @use '../../../00-tokens/tokens';
-@use '../../controls/text-link/text-link';
+@use '../../controls/text-link/yds-text-link';
 
 h1,
 h2,
