+++ components/02-molecules/accordion/_accordion-item.twig
@@ -24,7 +24,7 @@
 {% endset %}
 
 {% set accordion__item__heading__content %}
-  {% include "@atoms/controls/base/control.twig" with {
+  {% include "@atoms/controls/base/yds-control.twig" with {
     control__content: accordion__item__cta__content,
     control__base_class: 'toggle',
     control__blockname: accordion__item__base_class,
