+++ components/02-molecules/accordion/accordion.twig
@@ -46,7 +46,7 @@
       {% block list__content %}
         {% embed "@atoms/lists/_list-item.twig" %}
           {% block list__item__content %}
-            {% include "@atoms/controls/base/control.twig" with {
+            {% include "@atoms/controls/base/yds-control.twig" with {
               control__content: expand_all__content,
               control__base_class: 'toggle-all',
               control__modifiers: ['expand'],
@@ -56,7 +56,7 @@
         {% endembed %}
         {% embed "@atoms/lists/_list-item.twig" %}
           {% block list__item__content %}
-            {% include "@atoms/controls/base/control.twig" with {
+            {% include "@atoms/controls/base/yds-control.twig" with {
               control__content: collapse_all__content,
               control__base_class: 'toggle-all',
               control__modifiers: ['collapse'],
