+++ components/02-molecules/alert/alert.stories.js
@@ -1,6 +1,6 @@
 import alertTwig from './alert.twig';
 import textFieldTwig from '../text/text-field.twig';
-import ctaTwig from '../../01-atoms/controls/cta/cta.twig';
+import ctaTwig from '../../01-atoms/controls/cta/yds-cta.twig';
 
 import alertData from './alert.yml';
