+++ components/02-molecules/menu/menu-toggle/menu-toggle.twig
@@ -9,7 +9,7 @@
   </span>
 {% endset %}
 
-{% include "@atoms/controls/base/control.twig" with {
+{% include "@atoms/controls/base/yds-control.twig" with {
   control__base_class: main_menu_toggle__base_class,
   control__blockname: main_menu_toggle__blockname,
   aria_expanded: 'false',
