+++ components/02-molecules/social-links/_social-link__content.twig
@@ -9,7 +9,7 @@
   } %}
 {% endset %}
 
-{% include "@atoms/controls/base/control.twig" with {
+{% include "@atoms/controls/base/yds-control.twig" with {
   control__blockname: social_links__baseclass,
   control__url: icon.url,
   control__content: social_link__content
