+++ components/02-molecules/text-with-image/text-with-image.twig
@@ -58,7 +58,7 @@
       } %}
       {# Link #}
       {% if text_with_image__link__content and text_with_image__link__url %}
-        {% include "@atoms/controls/text-link/text-link.twig" with {
+        {% include "@atoms/controls/text-link/yds-text-link.twig" with {
           link__content: text_with_image__link__content,
           link__url: text_with_image__link__url,
         } %}
