+++ components/03-organisms/menu/breadcrumbs/breadcrumbs.twig
@@ -22,7 +22,7 @@
     } %}
     <span aria-hidden="true">{{ parent_label }}</span>
   {% endset %}
-  {% include "@atoms/controls/base/control.twig" with {
+  {% include "@atoms/controls/base/yds-control.twig" with {
     control__blockname: breadcrumbs__base_class,
     aria_expanded: 'false',
     aria_controls: 'breadcrumbs-inner',
