+++ components/03-organisms/menu/utility-nav/utility-nav.twig
@@ -21,7 +21,7 @@
   {% endif %}
   {# Link #}
   {% if utility_nav__link__content and utility_nav__link__url %}
-    {% include "@atoms/controls/cta/cta.twig" with {
+    {% include "@atoms/controls/cta/yds-cta.twig" with {
       cta__blockname: utility_nav__base_class,
       cta__content: utility_nav__link__content,
       cta__href: utility_nav__link__url,
