+++ components/03-organisms/site-footer/site-footer.twig
@@ -55,13 +55,13 @@
       </div>
       {# Meta #}
       <div {{ bem('meta', [], site_footer__base_class) }}>
-        {% include "@atoms/controls/text-link/text-link.twig" with {
+        {% include "@atoms/controls/text-link/yds-text-link.twig" with {
           link__content: 'Accessibility at Yale',
           link__url: "https://usability.yale.edu/web-accessibility/accessibility-yale",
           link__style: ['no-underline'],
         } %}
         <span {{ bem('divider', [], site_footer__base_class) }}>|</span>
-        {% include "@atoms/controls/text-link/text-link.twig" with {
+        {% include "@atoms/controls/text-link/yds-text-link.twig" with {
           link__content: 'Privacy Policy',
           link__url: "https://www.yale.edu/privacy-policy",
           link__style: ['no-underline'],
