+++ components/03-organisms/site-header/site-header.twig
@@ -27,7 +27,7 @@
 {% set site_header__primary %}
   <div {{ bem('primary', [], site_header__base_class) }}>
     {# Site Branding #}
-    {% include "@atoms/controls/text-link/text-link.twig" with {
+    {% include "@atoms/controls/text-link/yds-text-link.twig" with {
       link__content: site_name,
       link__url: '/',
       link__base_class: 'site-branding',
@@ -47,7 +47,7 @@
 {% set site_header__secondary %}
   <div {{ bem('secondary', [], site_header__base_class) }}>
     {# Yale Branding #}
-    {% include "@atoms/controls/text-link/text-link.twig" with {
+    {% include "@atoms/controls/text-link/yds-text-link.twig" with {
       link__content: 'Yale University',
       link__url: 'https://www.yale.edu',
       link__base_class: 'yale-branding',
@@ -66,7 +66,7 @@
   <div {{ bem('inner', [], site_header__base_class) }}>
     <div {{ bem('mobile-header', [], site_header__base_class) }}>
       {# Yale Branding #}
-      {% include "@atoms/controls/text-link/text-link.twig" with {
+      {% include "@atoms/controls/text-link/yds-text-link.twig" with {
         link__content: site_name,
         link__url: '/',
         link__base_class: 'site-branding',
